moved component to common, added comment about why local copy is needed
